feat(accountpool): 账号用量暂停阈值改为三层继承，拦截按快照实时判定并在阈值变更后重算运行态
全局阈值调高后，旧阈值写下的 Redis usage_pause 标记会把账号继续锁到窗口重置，而被暂停的账号拿不到
流量、也就没有新观测去触发恢复。改为：

- 阈值按「账号 → 池型渠道 → 全局」三层继承（core/accountusage）：账号与渠道两层 NULL 继承、1~100 覆写，
  不接受 0；迁移 000085 新增两列与 CHECK（渠道列 pool-only）。
- 候选准备按账号 usage_snapshot 水位与生效阈值实时派生用量暂停，覆盖 Redis 标记；阈值改动对下一次请求即生效。
- Recorder 观测时按账号解析生效阈值，快照落库把相对重置秒数换成绝对 reset_at。
- 新增 health.Reconciler：全局 setting 写入钩子 / 渠道阈值变更 / 账号阈值独立接口
  （PUT /subscription-accounts/{id}/usage-pause-threshold）保存后分别重算全部 / 该渠道 / 该账号的标记，
  结果以 runtime_refresh 回给管理端；账号视图新增生效阈值与来源。

+++ internal/bootstrap/worker_server.go
@@ -155,7 +155,7 @@ func NewWorkerServerApp(ctx context.Context, deps WorkerServerAppDeps) (*WorkerS
 	channelModelInventoryService.WithAccountResolver(probeIdentity)
 	// 验证成功回填账号观测（水位 + LRU），与请求路径同一 Recorder。
 	probeHealth := subscriptionhealth.NewRecorder(queries, permitStore, deps.Logger, 0).
-		WithThresholdProvider(func(ctx context.Context) float64 {
+		WithThresholdProvider(func(ctx context.Context) int32 {
 			return appsettings.GatewayAccountUsagePauseThreshold(ctx, settingsStore)
 		})
 	channelTestService.WithAccountHealth(probeHealth)
